+++ quill_native_bridge_platform_interface/lib/src/platform_feature.dart
@@ -9,5 +9,6 @@ enum QuillNativeBridgeFeature {
   getClipboardFiles,
   openGalleryApp,
   saveImageToGallery,
-  saveImage;
+  saveImage,
+  isAppleSafari;
 }